OCM-2180 | fix: hide available upgrades for HCP in list versions
The target is to transition to cluster based updates recommandations,
which means no more available upgrades for a single version, only for a specific cluster.
So the rosa list version --hosted-cp should not show anymore the available upgrades,
as in any case the current list is not correct (also showing versions which are disabled for HCP).
The user is informed that it should use rosa list upgrades -c clusterID instead