fix: invalid default value and interaction flow
The default value for the route selector for ingress option was causing
the interactive flow to fail, without giving the user a chance to input
a correct value after the first try. This commit fixes the check so that
the default value is valid, and adds a validator to the flow so that the
user can try again if invalid input was provided.